Create README.md
To create an animated calendar with HTML , CSS and JavaScript. First we will be using the table tag, which will be used to create the structure of the calendar. This will give us an idea of how the
calendar is created using HTML. Later we will apply some CSS property to make the design of the calendar better.